Add tests (#5)
* change material banner text

* create user circle avatar widget

* fix user avatar widget

* add user circle avatr widget to more pages

* remove comments

* calendar widget wip

* remove dependencies

* generate modal of tiles on calendar

* finish user stats layout

* raw calendar implementation needs refactor for performance

* add group extension on iterable

* add test on group by extension

* format code

* use group by method extension

* order by date

* create mock endpoint implementations for delete and update

* add patch and delete to mock router

* add delete log method to exerciseLog data source

* create update exercise log dto

* update exercise log data_source implementation

* fix wrong dto param

* add methods to exercise_log_repository

* edit check in page cubit wip

* remove imports

* create edit check in page arguments

* route setup on app router

* edit check in page setup

* register cubit in injector

* add behavior to navigate to edit page

* add initial state to form fields

* edit page wip

* add upload repository as a dependency of cubit

* update exercise log feature complete

* add CheckInPgeCubit in dependency injector

* delete check in feature

* setup router for edit_gym_challenge_page

* add navigation to edit challenge page

* edit challenge page skeleton

* add method to formating to slash separated date

* main structure of edit challenge page

* add date readonly

* edit gym challenge page wip

* add update gym challenge api call to data source

* add left zero to date helper

* add dto

* add update challenge to repository

* udpate gym challenge feature

* add modal to user calendars page

* fix date form in edit gym challenge

* fix font size

* remove placeholder image

* remove comments

* Refactor/flutter app performance improvements (#3)

* add devtools to track shared preferences

* change pushing routes to widget stack strategy

* change navigation method to pop previous pages

* remove previous widgets from the stack when navigating back

* remove previous widgets on navigate

* fix error with mocked data

* code formating

* change constant padding variable to double

* add enviroment variable config

* add environment variable to http client

* add archictecture documentation

* add tests

* cahnge logo

* add packages for testing

* add date helper tests

* remove http package

* test group code wip

* fix typo

* configure settings.gradle

* create validatable interface

* add validators

* create uuid service wrapper

* add validators

* add mocks to .gitignore

* add group_code_validator tests

* test email validator

* add tests for username validator

* add generated files to gitignore

* add test to constructor

* add title validation tests

* add more form validators

* Add unit testing(#4)

* refactor validators in creat account form

* refactor validators in sign in form

* switch to refactored validator implementation

* switch validator

* switch validator

* add test for abstract application exception

* add test for invalid_token_exception

* add tests exceptions

* add tests for auth data source

* add tests for auth_repository

* add test for logged user data source

* remove unused code

* add tests for logged user data source

* test exercise log data source

* remove unused method

* remove unnecessary try catch blocks

* add tests for exercise log data source

* add widgets tests

* move date helper method

* use helper from DateHelper class

* add test for upload repository

* create non static memory cache class

* memoryc cache test service wip

* add fake async lib for testing time

* test memory cache service

* add cache service as a dependency to repositories

* add CacheService as a dependency to repositories

* remove unused methods

* ignore memory implementations of data sources for test coverage

* add test for logout  in auth_repository

* unit test for http client in progress

* format project code

* add error handling test to dio http client

* add tests for methods in http client

* add test for application exception

* add splash page test

* add welcome page test

+++ jim_hats_mobile/lib/presentation/views/create-chalenge/create_challenge_page.dart
@@ -93,8 +93,8 @@ class _CreateChallengeViewState extends State<CreateChallengeView> {
       ),
       body: SafeArea(
           child: Padding(
-        padding: EdgeInsets.symmetric(
-            horizontal: AppSpacings.horizontalPadding.toDouble()),
+        padding:
+            EdgeInsets.symmetric(horizontal: AppSpacings.horizontalPadding),
         child: Form(
           key: _formKey,
           child: ListView(
